+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
 ## Unreleased
+- Supported server performance breakdown by operation name
+  ([#731](https://github.com/census-instrumentation/opencensus-python/pull/731))
 
 ## 0.3.1
+Released 2019-06-30
 - Added metrics exporter
   ([#678](https://github.com/census-instrumentation/opencensus-python/pull/678))