+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## Unreleased
+- Added standard metrics
+  ([#708](https://github.com/census-instrumentation/opencensus-python/pull/708),
+   [#718](https://github.com/census-instrumentation/opencensus-python/pull/718),
+   [#720](https://github.com/census-instrumentation/opencensus-python/pull/720),
+   [#722](https://github.com/census-instrumentation/opencensus-python/pull/722),
+   [#724](https://github.com/census-instrumentation/opencensus-python/pull/724))
 
 ## 0.3.1
 - Added metrics exporter