+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -7,6 +7,8 @@
    [#720](https://github.com/census-instrumentation/opencensus-python/pull/720),
    [#722](https://github.com/census-instrumentation/opencensus-python/pull/722),
    [#724](https://github.com/census-instrumentation/opencensus-python/pull/724))
+- Supported server performance breakdown by operation name
+  ([#735](https://github.com/census-instrumentation/opencensus-python/pull/735))
 
 ## 0.3.1
 - Added metrics exporter