+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased
+
+## 0.7.0
+Released 2019-07-31
+
 - Added standard metrics
   ([#708](https://github.com/census-instrumentation/opencensus-python/pull/708),
    [#718](https://github.com/census-instrumentation/opencensus-python/pull/718),
@@ -11,16 +15,20 @@
   ([#735](https://github.com/census-instrumentation/opencensus-python/pull/735))
 
 ## 0.3.1
+Released 2019-06-30
+
 - Added metrics exporter
   ([#678](https://github.com/census-instrumentation/opencensus-python/pull/678))
 
 ## 0.2.1
 Released 2019-06-13
+
 - Support span attributes
   ([#682](https://github.com/census-instrumentation/opencensus-python/pull/682))
 
 ## 0.2.0
 Released 2019-05-31
+
 - Added log exporter
   ([#657](https://github.com/census-instrumentation/opencensus-python/pull/657),
   [#668](https://github.com/census-instrumentation/opencensus-python/pull/668))